refactor: 테스트코드의 패키지 위치는 프로덕션코드의 위치와 동일하게 맞추기
