+++ src/test/java/racingcar/controller/GameControllerTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
