+++ src/test/java/racingcar/utils/ParserTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
